# api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/discounts": {
      "post": {
        "operationId": "DiscountsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDiscountDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Code promo créé avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscountResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Code promo déjà existant"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Créer un code promo (Admin uniquement)",
        "tags": [
          "discounts"
        ]
      },
      "get": {
        "operationId": "DiscountsController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Numéro de page",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Nombre d'éléments par page",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "Filtrer par statut actif",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Liste des codes promo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedDiscountsResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Récupérer tous les codes promo (Admin uniquement)",
        "tags": [
          "discounts"
        ]
      }
    },
    "/discounts/stats": {
      "get": {
        "operationId": "DiscountsController_getStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Statistiques"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Statistiques des codes promo (Admin uniquement)",
        "tags": [
          "discounts"
        ]
      }
    },
    "/discounts/{id}": {
      "get": {
        "operationId": "DiscountsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID du code promo",
            "schema": {
              "example": "uuid-discount",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Détails du code promo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscountResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Code promo non trouvé"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Récupérer un code promo par ID (Admin uniquement)",
        "tags": [
          "discounts"
        ]
      },
      "patch": {
        "operationId": "DiscountsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID du code promo",
            "schema": {
              "example": "uuid-discount",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDiscountDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Code promo mis à jour",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscountResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Code promo non trouvé"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Mettre à jour un code promo (Admin uniquement)",
        "tags": [
          "discounts"
        ]
      },
      "delete": {
        "operationId": "DiscountsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID du code promo",
            "schema": {
              "example": "uuid-discount",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Code promo supprimé"
          },
          "404": {
            "description": "Code promo non trouvé"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Supprimer un code promo (Admin uniquement)",
        "tags": [
          "discounts"
        ]
      }
    },
    "/discounts/{id}/activate": {
      "patch": {
        "operationId": "DiscountsController_activate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID du code promo",
            "schema": {
              "example": "uuid-discount",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Code promo activé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscountResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Activer un code promo (Admin uniquement)",
        "tags": [
          "discounts"
        ]
      }
    },
    "/discounts/{id}/deactivate": {
      "patch": {
        "operationId": "DiscountsController_deactivate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID du code promo",
            "schema": {
              "example": "uuid-discount",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Code promo désactivé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscountResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Désactiver un code promo (Admin uniquement)",
        "tags": [
          "discounts"
        ]
      }
    },
    "/discounts/validate": {
      "post": {
        "description": "Vérifie si un code promo est valide et utilisable",
        "operationId": "DiscountsController_validate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateDiscountDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Résultat de la validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiscountValidationDto"
                }
              }
            }
          }
        },
        "summary": "Valider un code promo",
        "tags": [
          "discounts"
        ]
      }
    },
    "/discounts/apply": {
      "post": {
        "description": "Calcule la remise pour un montant donné",
        "operationId": "DiscountsController_apply",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyDiscountDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Résultat de l'application",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplyDiscountResultDto"
                }
              }
            }
          },
          "400": {
            "description": "Code promo invalide ou conditions non remplies"
          }
        },
        "summary": "Appliquer un code promo",
        "tags": [
          "discounts"
        ]
      }
    },
...
  },
  "info": {
    "title": "API Documentation",
    "description": "API documentation for the application",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      },
      "JWT-refresh": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Refresh-JWT",
        "description": "Enter refresh JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateDiscountDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Code promo unique",
            "example": "SUMMER2024",
            "maxLength": 50
          },
          "type": {
            "type": "string",
            "description": "Type de remise (PERCENTAGE ou FIXED)",
            "enum": [
              "PERCENTAGE",
              "FIXED"
            ],
            "example": "PERCENTAGE"
          },
          "value": {
            "type": "number",
            "description": "Valeur de la remise (pourcentage ou montant fixe en TND)",
            "example": 20,
            "minimum": 0
          },
          "description": {
            "type": "string",
            "description": "Description de la promotion",
            "example": "Réduction été 2024",
            "maxLength": 500
          },
          "expiresAt": {
            "type": "string",
            "description": "Date d'expiration du code promo",
            "example": "2024-12-31T23:59:59.000Z"
          },
          "minAmount": {
            "type": "number",
            "description": "Montant minimum pour appliquer le code (en TND)",
            "example": 50,
            "minimum": 0
          },
          "maxUses": {
            "type": "number",
            "description": "Nombre maximum d'utilisations",
            "example": 100,
            "minimum": 1
          },
          "isActive": {
            "type": "boolean",
            "description": "Code actif ou non",
            "default": true
          }
        },
        "required": [
          "code",
          "type",
          "value",
          "expiresAt"
        ]
      },
      "DiscountResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID unique du code promo",
            "example": "uuid-discount"
          },
          "code": {
            "type": "string",
            "description": "Code promo",
            "example": "SUMMER2024"
          },
          "type": {
            "type": "string",
            "description": "Type de remise",
            "enum": [
              "PERCENTAGE",
              "FIXED"
            ],
            "example": "PERCENTAGE"
          },
          "value": {
            "type": "number",
            "description": "Valeur de la remise",
            "example": 20
          },
          "description": {
            "type": "string",
            "description": "Description de la promotion",
            "example": "Réduction été 2024"
          },
          "isActive": {
            "type": "boolean",
            "description": "Code actif ou non",
            "example": true
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string",
            "description": "Date d'expiration",
            "example": "2024-12-31T23:59:59.000Z"
          },
          "minAmount": {
            "type": "number",
            "description": "Montant minimum pour appliquer le code",
            "example": 50
          },
          "maxUses": {
            "type": "number",
            "description": "Nombre maximum d'utilisations",
            "example": 100
          },
          "usedCount": {
            "type": "number",
            "description": "Nombre d'utilisations actuelles",
            "example": 25
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Date de création",
            "example": "2024-01-15T10:30:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Date de dernière mise à jour",
            "example": "2024-01-15T10:30:00.000Z"
          }
        },
        "required": [
          "id",
          "code",
          "type",
          "value",
          "isActive",
          "expiresAt",
          "usedCount",
          "createdAt",
          "updatedAt"
        ]
      },
      "PaginatedDiscountsResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "description": "Liste des codes promo",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DiscountResponseDto"
            }
          },
          "meta": {
            "type": "object",
            "description": "Métadonnées de pagination",
            "example": {
              "total": 50,
              "page": 1,
              "limit": 10,
              "totalPages": 5
            }
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "ValidateDiscountDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Code promo à valider",
            "example": "SUMMER2024"
          }
        },
        "required": [
          "code"
        ]
      },
      "DiscountValidationDto": {
        "type": "object",
        "properties": {
          "isValid": {
            "type": "boolean",
            "description": "Code valide ou non",
            "example": true
          },
          "code": {
            "type": "string",
            "description": "Code promo",
            "example": "SUMMER2024"
          },
          "message": {
            "type": "string",
            "description": "Message explicatif",
            "example": "Code promo valide"
          },
          "discount": {
            "description": "Détails du code promo si valide",
            "allOf": [
              {
                "$ref": "#/components/schemas/DiscountResponseDto"
              }
            ]
          }
        },
        "required": [
          "isValid",
          "code"
        ]
      },
      "ApplyDiscountDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Code promo à appliquer",
            "example": "SUMMER2024"
          },
          "amount": {
            "type": "number",
            "description": "Montant total de la commande (en TND)",
            "example": 150,
            "minimum": 0
          }
        },
        "required": [
          "code",
          "amount"
        ]
      },
      "ApplyDiscountResultDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Code promo appliqué",
            "example": "SUMMER2024"
          },
          "type": {
            "type": "string",
            "description": "Type de remise",
            "enum": [
              "PERCENTAGE",
              "FIXED"
            ],
            "example": "PERCENTAGE"
          },
          "value": {
            "type": "number",
            "description": "Valeur de la remise",
            "example": 20
          },
          "originalAmount": {
            "type": "number",
            "description": "Montant original (avant remise)",
            "example": 150
          },
          "discountAmount": {
            "type": "number",
            "description": "Montant de la remise",
            "example": 30
          },
          "finalAmount": {
            "type": "number",
            "description": "Montant final (après remise)",
            "example": 120
          },
          "message": {
            "type": "string",
            "description": "Message de confirmation",
            "example": "Code promo appliqué avec succès ! Vous économisez 30 TND"
          }
        },
        "required": [
          "code",
          "type",
          "value",
          "originalAmount",
          "discountAmount",
          "finalAmount",
          "message"
        ]
      },
...
    }
  }
}